refactor: seperate handler and writer
